DHFPROD-9751: Convert services and transform to JavaScript Module (mjs)
