+++ marklogic-data-hub/src/main/resources/ml-modules/services/hubFlowRunner.mjs
@@ -15,11 +15,11 @@
  */
 'use strict';
 
-const flowApi = require("/data-hub/public/flow/flow-api.sjs");
-const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
-const json = require('/MarkLogic/json/json.xqy');
+import flowApi from "/data-hub/public/flow/flow-api.mjs";
+import httpUtils from "/data-hub/5/impl/http-utils.mjs";
+import json from "/MarkLogic/json/json.xqy";
 
-function post(context, params, input) {
+export function post(context, params, input) {
   xdmp.securityAssert("http://marklogic.com/data-hub/privileges/run-step", "execute");
   return isXmlInput(context) ? processXmlInput(input) : processJsonInput(input);
 }
@@ -35,7 +35,7 @@ function isXmlInput(context) {
 function processJsonInput(input) {
   input = input.toObject();
   const jobId = input.jobId || sem.uuidString();
-  
+
   let stepNumbers = input.steps || [];
   if (stepNumbers && !Array.isArray(stepNumbers)) {
     stepNumbers = [stepNumbers];
@@ -104,5 +104,3 @@ function buildContentArray(xmlInput) {
   }
   return contentArray;
 }
-
-exports.POST = post;
