+++ marklogic-data-hub/src/main/resources/ml-modules/services/mlRunFlow.mjs
@@ -14,8 +14,8 @@
  limitations under the License.
  */
 'use strict';
-const DataHubSingleton = require("/data-hub/5/datahub-singleton.sjs");
-const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
+import DataHubSingleton from "/data-hub/5/datahub-singleton.mjs";
+import httpUtils from "/data-hub/5/impl/http-utils.mjs";
 
 /**
  * DO NOT USE THIS; it is deprecated as of DHF 5.3.0.
@@ -23,11 +23,11 @@ const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
  * Use the processBatch.mjs DS endpoint in StepRunnerService instead.
  */
 
-function get(context, params) {
+export function get(context, params) {
   return post(context, params, null);
 }
 
-function post(context, params, input) {
+export function post(context, params, input) {
   let flowName = params["flow-name"];
   if (!fn.exists(flowName)) {
     httpUtils.throwBadRequestWithArray(["Bad Request", "Invalid request - must specify a flowName"]);
@@ -43,13 +43,8 @@ function post(context, params, input) {
   }
 }
 
-function put(context, params, input) {
+export function put(context, params, input) {
 }
 
-function deleteFunction(context, params) {
+export function deleteFunction(context, params) {
 }
-
-exports.GET = get;
-exports.POST = post;
-exports.PUT = put;
-exports.DELETE = deleteFunction;
