+++ marklogic-data-hub/src/main/resources/ml-modules/transforms/mlRunIngest.mjs
@@ -1,10 +1,10 @@
 'use strict';
-const DataHubSingleton = require("/data-hub/5/datahub-singleton.sjs");
+import DataHubSingleton from "/data-hub/5/datahub-singleton.mjs";
 const datahub = DataHubSingleton.instance();
-const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
-const provLib = require("/data-hub/5/impl/prov.sjs");
+import httpUtils from "/data-hub/5/impl/http-utils.mjs";
+import provLib from "/data-hub/5/impl/prov.mjs";
 
-function transform(context, params, content) {
+export function transform(context, params, content) {
   let flowName = params['flow-name'] ? xdmp.urlDecode(params['flow-name']) : "default-ingestion";
   let flow = datahub.flow.getFlow(flowName);
   if (!flow) {
@@ -68,5 +68,3 @@ function transform(context, params, content) {
    provLib.getProvenanceWriteQueue().persist();
    return Sequence.from(docs);
 }
-
-exports.transform = transform;
