+++ marklogic-data-hub/src/main/resources/ml-modules/transforms/mlSjsInputFlow.mjs
@@ -13,12 +13,12 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
-const consts = require("/data-hub/4/impl/consts.sjs");
-const flowlib = require("/data-hub/4/impl/flow-lib.sjs");
-const tracelib = require("/data-hub/4/impl/trace-lib.sjs");
-const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
+import consts from "/data-hub/4/impl/consts.mjs";
+import flowlib from "/data-hub/4/impl/flow-lib.mjs";
+import tracelib from "/data-hub/4/impl/trace-lib.mjs";
+import httpUtils from "/data-hub/5/impl/http-utils.mjs";
 
-function transform(context, params, content) {
+export function transform(context, params, content) {
   let jobId = params["job-id"] ||sem.uuidString();
   let entityName = params['entity-name'];
   let flowName = params['flow-name'];
@@ -45,5 +45,3 @@ function transform(context, params, content) {
 
   return envelope;
 }
-
-exports.transform = transform;
